requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
clearbit==0.1.7
cryptography==2.7
dataclasses==0.6
Django==2.2.4
django-rest-auth==0.9.5
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djongo==1.2.33
future==0.17.1
idna==2.8
mongoengine==0.18.2
pycparser==2.19
pyhunter==0.3
PyJWT==1.7.1
pymongo==3.9.0
pyTelegramBotAPI==3.6.6
pytz==2019.2
requests==2.7.0
six==1.12.0
sqlparse==0.2.4
urllib3==1.25.3
